Store PPE 400 programs in a sectioned container
Runtime 400 still used the PCBoard container. Its 16 bit counts, byte wide
type ids and 16 bit VAR mask capped code size, declarations, records and
routine parameters, and the constructs decided in S1 and S2 had no on-disk
encoding at all: record layouts with host or dynamic fields and short
circuit operators were rejected when writing a file, so those programs
could only run from memory. Host types were bound by their stored numeric
id and enums by their complete stored value list, which made extending the
API a compatibility risk for existing files.

Runtime 400 now writes its own container, recognised by a magic number
rather than a version field. It has a header, a section directory with
explicit offsets and lengths, 32 bit ids, per parameter VAR modes, typed
length framed UTF-8 and binary constants, optional Zstd per section,
optional debug names, a SHA-256 content identity and a loader that
validates the whole file before the VM runs an instruction. Host imports
are bound by qualified name and expected signature, so the catalog may be
renumbered and extended without invalidating old files. Container,
bytecode, section schema and host ABI are versioned separately; unknown
optional sections are skipped and unknown required data is refused.

Pre-400 PPEs keep their container, encryption and execution semantics.
PPEs from the unreleased 400 beta use the old container and are refused
with a request to recompile.

Three defects surfaced while connecting the new path. Control commands
such as STOP and RETURN were only turned into their own instructions by
the legacy decoder, which the new path does not use, breaking error
handling and cleanup; they are now normalised when they are generated.
Record I/O wrote DATE and EDATE as 16 bit but read them back into a 32 bit
value, leaving the high bits undefined. pplc --disassemble panicked on the
default runtime because it still serialised instructions into the dropped
word encoding.

Verified with the five PPL packages in separate English and German
processes: 2902 passed, 0 failed, 6 ignored in each. The all-target check
of those packages and the engine without default features pass.

Known gaps, also listed in the release plan: META is reserved but carries
no meaning, and unknown optional sections are lost when a file is
rewritten; debug data holds variable names only; the compiler still
budgets code in legacy units; record field bounds remain 16 bit
internally; Zstd is built and tested on Linux only; the loader has not
been fuzzed.

+++ crates/icy_board_engine/src/compiler/user_data.rs
@@ -44,7 +44,7 @@ pub trait UserDataValue: Send + Sync {
 }
 
 #[cfg(feature = "bbs")]
-pub(crate) fn runtime_object(object: &(dyn std::any::Any + Send + Sync), type_id: u8) -> crate::Res<&dyn UserDataValue> {
+pub(crate) fn runtime_object(object: &(dyn std::any::Any + Send + Sync), type_id: u32) -> crate::Res<&dyn UserDataValue> {
     let object = object.downcast_ref::<ResourceUserData>().map_or(object, |resource| resource.object.as_ref());
     object
         .downcast_ref::<RuntimeUserData>()
