+++ crates/icy_board_engine/src/icy_board/state/ppl_board.rs
@@ -57,8 +57,8 @@ impl UserData for PplBoard {
     const TYPE_NAME: &'static str = "Board";
     const EMPTY_VALUE: Option<fn() -> VariableValue> = Some(|| {
         Self {
-            conferences: VariableValue::new_vector(crate::executable::VariableType::UserData(crate::parser::CONFERENCE_ID as u8), Vec::new()),
-            users: VariableValue::new_vector(crate::executable::VariableType::UserData(crate::parser::USER_ID as u8), Vec::new()),
+            conferences: VariableValue::new_vector(crate::executable::VariableType::UserData(crate::parser::CONFERENCE_ID as u32), Vec::new()),
+            users: VariableValue::new_vector(crate::executable::VariableType::UserData(crate::parser::USER_ID as u32), Vec::new()),
             ..Self::default()
         }
         .value()
